requirements.txt
Django==4.2.20
python-telegram-bot==13.15
environs==14.1.1